+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added schema versioning, and an automatic schema migration system, including datetime normalization.
 - Rename old `stats` command to `list`
 - Revamped the old `stats` command to display overall stats for collected data.
+- Added `--detailed` flag for `activity` and `sleep` commands.
 
 ## [1.0.0]